+++ liberty-maven-plugin/src/main/java/net/wasdev/wlp/maven/plugins/applications/LooseEarApplication.java
@@ -1,28 +1,13 @@
 package net.wasdev.wlp.maven.plugins.applications;
 
-import java.io.File;
-import java.io.FileOutputStream;
-import java.util.List;
-import java.util.jar.Attributes;
-import java.util.jar.Manifest;
-
 import org.apache.maven.artifact.Artifact;
-import org.apache.maven.model.Resource;
 import org.apache.maven.project.MavenProject;
-import org.codehaus.plexus.util.xml.Xpp3Dom;
 import org.w3c.dom.Element;
 
-public class LooseEarApplication {
-    private MavenProject project;
-    private LooseConfigData config;
+public class LooseEarApplication extends LooseApplication {
     
     public LooseEarApplication(MavenProject project, LooseConfigData config) {
-        this.project = project;
-        this.config = config;
-    }
-    
-    public LooseConfigData getConfig() {
-        return config;
+        super(project, config);
     }
     
     public Element addJarModule(MavenProject proj) throws Exception {
@@ -33,11 +18,7 @@ public class LooseEarApplication {
         }
         Element jarArchive = config.addArchive(jarArchiveName);
         config.addDir(jarArchive, proj.getBuild().getOutputDirectory(), "/");
-        @SuppressWarnings("unchecked")
-        List<Resource> resources = proj.getResources();
-        for (Resource res : resources) {
-            config.addDir(jarArchive, res.getDirectory(), "/");
-        }
+        addResourceDir(jarArchive, proj, "/");
         // add manifest.mf
         if ("ejb".equals(proj.getPackaging())) {
             addManifestFile(jarArchive, proj, "maven-ejb-plugin");
@@ -53,22 +34,10 @@ public class LooseEarApplication {
         Element warArchive = config.addArchive(warArchiveName);
         config.addDir(warArchive, warSourceDir, "/");
         config.addDir(warArchive, proj.getBuild().getOutputDirectory(), "/WEB-INF/classes");
-        @SuppressWarnings("unchecked")
-        List<Resource> resources = proj.getResources();
-        for (Resource res : resources) {
-            config.addDir(warArchive, res.getDirectory(), "/WEB-INF/classes");
-        }
+        addResourceDir(warArchive, proj, "/WEB-INF/classes");
         // add Manifest file
         addManifestFile(warArchive, proj, "maven-war-plugin");
         return warArchive;
-    }
-    
-    public void addManifestFile(Element e, MavenProject proj, String pluginId) throws Exception {
-        config.addFile(e, getManifestFile(proj, "org.apache.maven.plugins", pluginId), "/META-INF/MANIFEST.MF");    
-    }
-    
-    public void addManifestFile(MavenProject proj, String pluginId) throws Exception {
-        config.addFile(getManifestFile(proj, "org.apache.maven.plugins", pluginId), "/META-INF/MANIFEST.MF");    
     }
     
     public void addModuleFromM2(Artifact artifact, String artifactFile) {
@@ -125,47 +94,5 @@ public class LooseEarApplication {
     
     public String getEarDefaultLibBundleDir() {
         return getPluginConfiguration("org.apache.maven.plugins", "maven-ear-plugin", "defaultLibBundleDir");
-    }
-    
-    public String getPluginConfiguration(String pluginGroupId, String pluginArtifactId, String key) {
-        Xpp3Dom dom = project.getGoalConfiguration(pluginGroupId, pluginArtifactId, null, null);
-        if (dom != null) {
-            Xpp3Dom val = dom.getChild(key);
-            if (val != null) {
-                return val.getValue();
-            }
-        }
-        return null;
-    }
-    
-    public String getManifestFile(MavenProject proj, String pluginGroupId, String pluginArtifactId) throws Exception {
-        Xpp3Dom dom = proj.getGoalConfiguration(pluginGroupId, pluginArtifactId, null, null);
-        if (dom != null) {
-            Xpp3Dom archive = dom.getChild("archive");
-            if (archive != null) {
-                Xpp3Dom val = archive.getChild("manifestFile");
-                if (val != null) {
-                    return proj.getBasedir().getAbsolutePath() + "/" + val.getValue();
-                }
-            }
-        }
-        return getDefaultManifest().getCanonicalPath();
-    }
-    
-    private File defaultMF = null;
-    
-    public File getDefaultManifest() throws Exception {
-        if (defaultMF == null) {
-            defaultMF = new File(
-                    project.getBuild().getDirectory() + "/liberty-maven/resources/META-INF/MANIFEST.MF");
-            defaultMF.getParentFile().mkdirs();
-            FileOutputStream fos = new FileOutputStream(defaultMF);
-            
-            Manifest manifest = new Manifest();
-            manifest.getMainAttributes().put(Attributes.Name.MANIFEST_VERSION, "1.0");
-            manifest.write(fos);
-            fos.close();
-        }
-        return defaultMF;
     }
 }
