+++ liberty-maven-plugin/src/main/java/net/wasdev/wlp/maven/plugins/applications/LooseWarApplication.java
@@ -1,0 +1,10 @@
+package net.wasdev.wlp.maven.plugins.applications;
+
+import org.apache.maven.project.MavenProject;
+
+public class LooseWarApplication extends LooseApplication {
+    
+    public LooseWarApplication(MavenProject project, LooseConfigData config) {
+        super(project, config);
+    }
+}
